Move demo to own folder and add support for code blocks
