+++ demo/Layout.jsx
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+
+export const Main = styled.main`
+  box-sizing: border-box;
+  display: grid;
+  height: 100vh;
+  margin: 0;
+  padding: 0;
+`;
+
+export const Container = styled.div`
+  margin: auto;
+  padding: 1em;
+  background-color: #fff;
+  color: #333;
+  border: 1px solid #f0f0f0;
+  border-radius: 5px;
+  text-align: center;
+  box-shadow: 0 2px 4px #00000018;
+
+  @media (max-width: 460px) {
+    width: 100%;
+    border-radius: 0;
+  }
+`;
