+++ demo/index.html
@@ -6,7 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
 
-    <title>Day and Time Picker [demo]</title>
+    <title>Day and Time Picker</title>
 
     <style>
       html {
@@ -31,6 +31,6 @@
   <body>
     <div id="demo-app"></div>
 
-    <script src="./demo.js"></script>
+    <script src="./index.js"></script>
   </body>
 </html>
